Add API tokens to our users
- Update our User model to automatically generate a token for every new user.
  and for current first user manually generate the token in rc

The Problem: Website Logins vs. API Keys
This is happening because a standard website login (which uses browser cookies and sessions) does not work for an API.
Think of it like this:
    our website login is a keycard that opens the front door of our application.
    An API is like a secure loading dock in the back. our front door keycard doesn't work there; we need a special, secret API Key to get access.
We need to create these secret API keys for our users and teach our API how to use them.